refactor: split config-commands.ts and core-commands.ts into one file per command
