+++ assistant/src/cli/memory.ts
@@ -1,13 +1,6 @@
 import type { Command } from "commander";
 
 import {
-  API_KEY_PROVIDERS,
-  getNestedValue,
-  loadRawConfig,
-  saveRawConfig,
-  setNestedValue,
-} from "../config/loader.js";
-import {
   dismissPendingConflicts,
   getMemorySystemStatus,
   queryMemory,
@@ -18,246 +11,11 @@ import {
 import { listPendingConflictDetails } from "../memory/conflict-store.js";
 import { listConversations } from "../memory/conversation-store.js";
 import { initializeDb, rawGet } from "../memory/db.js";
-import {
-  clearAllRules,
-  getAllRules,
-  removeRule,
-} from "../permissions/trust-store.js";
-import {
-  deleteSecureKey,
-  getSecureKey,
-  setSecureKey,
-} from "../security/secure-keys.js";
 import { getCliLogger } from "../util/logger.js";
 
 const log = getCliLogger("cli");
 
 const SHORT_HASH_LENGTH = 8;
-
-export function registerConfigCommand(program: Command): void {
-  const config = program.command("config").description("Manage configuration");
-
-  config
-    .command("set <key> <value>")
-    .description(
-      "Set a config value (supports dotted paths like apiKeys.anthropic)",
-    )
-    .action((key: string, value: string) => {
-      const raw = loadRawConfig();
-      // Try to parse as JSON for booleans/numbers, fall back to string
-      let parsed: unknown = value;
-      try {
-        parsed = JSON.parse(value);
-      } catch {
-        // keep as string
-      }
-      setNestedValue(raw, key, parsed);
-      saveRawConfig(raw);
-      log.info(`Set ${key} = ${JSON.stringify(parsed)}`);
-    });
-
-  config
-    .command("get <key>")
-    .description("Get a config value (supports dotted paths)")
-    .action((key: string) => {
-      const raw = loadRawConfig();
-      const value = getNestedValue(raw, key);
-      if (value === undefined) {
-        log.info(`(not set)`);
-      } else {
-        log.info(
-          typeof value === "object"
-            ? JSON.stringify(value, null, 2)
-            : String(value),
-        );
-      }
-    });
-
-  config
-    .command("list")
-    .description("List all config values")
-    .action(() => {
-      const raw = loadRawConfig();
-      if (Object.keys(raw).length === 0) {
-        log.info("No configuration set");
-      } else {
-        log.info(JSON.stringify(raw, null, 2));
-      }
-    });
-
-  config
-    .command("validate-allowlist")
-    .description("Validate regex patterns in secret-allowlist.json")
-    .action(() => {
-      const { validateAllowlistFile } =
-        // eslint-disable-next-line @typescript-eslint/no-require-imports
-        require("../security/secret-allowlist.js") as typeof import("../security/secret-allowlist.js");
-      try {
-        const errors = validateAllowlistFile();
-        if (errors == null) {
-          log.info("No secret-allowlist.json file found");
-          return;
-        }
-        if (errors.length === 0) {
-          log.info("All patterns in secret-allowlist.json are valid");
-          return;
-        }
-        log.error(
-          `Found ${errors.length} invalid pattern(s) in secret-allowlist.json:`,
-        );
-        for (const e of errors) {
-          log.error(`  [${e.index}] "${e.pattern}": ${e.message}`);
-        }
-        process.exit(1);
-      } catch (err) {
-        log.error(
-          `Failed to read secret-allowlist.json: ${(err as Error).message}`,
-        );
-        process.exit(1);
-      }
-    });
-}
-
-export function registerKeysCommand(program: Command): void {
-  const keys = program
-    .command("keys")
-    .description("Manage API keys in secure storage");
-
-  keys
-    .command("list")
-    .description("List all stored API key names")
-    .action(() => {
-      const stored: string[] = [];
-      for (const provider of API_KEY_PROVIDERS) {
-        const value = getSecureKey(provider);
-        if (value) stored.push(provider);
-      }
-      if (stored.length === 0) {
-        log.info("No API keys stored");
-      } else {
-        for (const name of stored) {
-          log.info(`  ${name}`);
-        }
-      }
-    });
-
-  keys
-    .command("set <provider> <key>")
-    .description("Store an API key (e.g. vellum keys set anthropic sk-ant-...)")
-    .action((provider: string, key: string) => {
-      if (setSecureKey(provider, key)) {
-        log.info(`Stored API key for "${provider}"`);
-      } else {
-        log.error(`Failed to store API key for "${provider}"`);
-        process.exit(1);
-      }
-    });
-
-  keys
-    .command("delete <provider>")
-    .description("Delete a stored API key")
-    .action((provider: string) => {
-      if (deleteSecureKey(provider)) {
-        log.info(`Deleted API key for "${provider}"`);
-      } else {
-        log.error(`No API key found for "${provider}"`);
-        process.exit(1);
-      }
-    });
-}
-
-export function registerTrustCommand(program: Command): void {
-  const trust = program.command("trust").description("Manage trust rules");
-
-  trust
-    .command("list")
-    .description("List all trust rules")
-    .action(() => {
-      const rules = getAllRules();
-      if (rules.length === 0) {
-        log.info("No trust rules");
-        return;
-      }
-      const idW = 8;
-      const toolW = 12;
-      const patternW = 30;
-      const scopeW = 20;
-      const decW = 6;
-      const priW = 4;
-      log.info(
-        "ID".padEnd(idW) +
-          "Tool".padEnd(toolW) +
-          "Pattern".padEnd(patternW) +
-          "Scope".padEnd(scopeW) +
-          "Dcn".padEnd(decW) +
-          "Pri".padEnd(priW) +
-          "Created",
-      );
-      log.info("-".repeat(idW + toolW + patternW + scopeW + decW + priW + 20));
-      for (const r of rules) {
-        const id = r.id.slice(0, SHORT_HASH_LENGTH);
-        const created = new Date(r.createdAt).toISOString().slice(0, 10);
-        log.info(
-          id.padEnd(idW) +
-            r.tool.padEnd(toolW) +
-            r.pattern.slice(0, patternW - 2).padEnd(patternW) +
-            r.scope.slice(0, scopeW - 2).padEnd(scopeW) +
-            r.decision.slice(0, decW - 1).padEnd(decW) +
-            String(r.priority).padEnd(priW) +
-            created,
-        );
-      }
-    });
-
-  trust
-    .command("remove <id>")
-    .description("Remove a trust rule by ID (or prefix)")
-    .action((id: string) => {
-      const rules = getAllRules();
-      const match = rules.find((r) => r.id.startsWith(id));
-      if (!match) {
-        log.error(`No rule found matching "${id}"`);
-        process.exit(1);
-      }
-      try {
-        removeRule(match.id);
-      } catch (err) {
-        log.error(err instanceof Error ? err.message : String(err));
-        process.exit(1);
-      }
-      log.info(
-        `Removed rule ${match.id.slice(0, SHORT_HASH_LENGTH)} (${match.tool}: ${
-          match.pattern
-        })`,
-      );
-    });
-
-  trust
-    .command("clear")
-    .description("Remove all trust rules")
-    .action(async () => {
-      const rules = getAllRules();
-      if (rules.length === 0) {
-        log.info("No trust rules to clear");
-        return;
-      }
-      const readline = await import("node:readline");
-      const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-      });
-      const answer = await new Promise<string>((resolve) => {
-        rl.question(`Remove all ${rules.length} trust rules? (y/N) `, resolve);
-      });
-      rl.close();
-      if (answer.toLowerCase() === "y") {
-        clearAllRules();
-        log.info(`Cleared ${rules.length} trust rules`);
-      } else {
-        log.info("Cancelled");
-      }
-    });
-}
 
 export function registerMemoryCommand(program: Command): void {
   const memory = program
